backend/requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
billiard==3.6.2.0
celery==4.4.0
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
vine==1.3.0
wrapt==1.11.2
zipp==2.2.0
